pyproject.toml
[project]
name = "pdf2md"
version = "0.1.0"
description = "Convert PDF files to Markdown using LightOnOCR-2-1B"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface_hub>=1.3.2",
    "logfire>=4.16.0",
    "pillow>=10.0.0",
    "pygments>=2.19.2",
    "pypdfium2>=4.0.0",
    "rich>=13.0.0",
]

[project.scripts]
pdf2md = "pdf2md.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
    "zensical",
    "ty>=0.0.12",
    "pre-commit>=4.5.1",
    "mkdocstrings>=1.0.1",
    "mkdocstrings-python>=2.0.1",
]

[project.optional-dependencies]
local = [
    "torch>=2.0.0",
    "torchvision>=0.24.1",
    "transformers @ git+https://github.com/huggingface/transformers",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py314"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = ["integration: requires Hugging Face credentials and endpoint access"]
